Move aar maven fix out of buildSrc
